pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "memprof_plotter"
version = "0.0.1"
dependencies = [
  "matplotlib",
  "PyGithub"
]
requires-python = ">=3.11"
description = "Plotting utility for tsp memory profiling data"

[tool.setuptools.packages.find]
include = ["memprof_plotter"]
exclude = ["memprof_plots"]

[project.scripts]
memprof_plotter = "memprof_plotter.plotter:main"